Integrated React Components

+++ src/controllers/ocrController.js
@@ -15,11 +15,11 @@ exports.getOCRResult = async (req, res) => {
     try {
         const doc = await db.collection('ocrData').doc(id).get();
         if (!doc.exists) {
-            return res.status(404).send('Document not found');
+            return res.status(404).json({message:'Document not found'});
         }
         res.status(200).json(doc.data());
     } catch (error) {
-        res.status(500).send('Error retrieving data from Firestore');
+        res.status(500).json({message:'Error retrieving data from Firestore'});
     }
 };
 
@@ -28,14 +28,14 @@ exports.updateOCRResult = async (req, res) => {
     const updatedData = req.body;
 
     if(id != updatedData.identification_number) {
-        res.status(400).send('Identification number do not match!!');
+        res.status(400).json({message:'Identification number do not match!!'});
     }
 
     try {
         await db.collection('ocrData').doc(id).set(updatedData, { merge: true });
-        res.status(200).send(`OCR data updated for ID: ${id}`);
+        res.status(200).json({message:`OCR data updated for ID: ${id}`});
     } catch (error) {
-        res.status(500).send('Error updating data in Firestore');
+        res.status(500).json({message:'Error updating data in Firestore'});
     }
 };
 
@@ -43,16 +43,20 @@ exports.deleteOCRResult = async (req, res) => {
     const id = req.params.id;
 
     try {
+        const doc = await db.collection('ocrData').doc(id).get();
+        if (!doc.exists) {
+            return res.status(404).json({message:'Document not found'});
+        }
         await db.collection('ocrData').doc(id).delete();
-        res.status(200).send(`Document with ID ${id} deleted`);
+        res.status(200).json({message:`Document with ID ${id} deleted`});
     } catch (error) {
-        res.status(500).send('Error deleting document from Firestore');
+        res.status(500).json({message:'Error deleting document from Firestore'});
     }
 };
 
 exports.createOCR = async (req, res) => {
     if (!req.file) {
-      return res.status(400).send('No file uploaded.');
+      return res.status(400).json({message:'No file uploaded.'});
     }
   
     
@@ -87,19 +91,16 @@ exports.createOCR = async (req, res) => {
         if(!file_name) {
             return res.status(400).send('Identification number not found in OCR data.');
         }
-        // Define output file path
-        const outputFilePath = path.join(outputDir, `${file_name}.json`);
-
-        // Check if the file already exists
-        if (fs.existsSync(outputFilePath)) {
-            return res.status(409).send(`A file with identification number ${file_name} already exists.`);
-        }
 
         try {
+            const doc = await db.collection('ocrData').doc(file_name).get();
+            if (doc.exists) {
+                return res.status(200).json({message:'ID already uploaded'});
+            }
             const docRef = await db.collection('ocrData').doc(OCRData.identification_number).set(OCRData);
-            res.status(200).send(`OCR data saved with ID: ${OCRData.identification_number}`);
+            res.status(200).json({message:`OCR data saved with ID: ${OCRData.identification_number}`});
         } catch (error) {
             console.error(error);
-            res.status(500).send('Error saving data to Firestore');
+            res.status(500).json({message:'Error saving data to Firestore'});
         }
   };
